证书管理框架基本完成，基本完成License生成的调用服务（包括幂等性设计），创建licenseResult数据表

+++ src/main/java/com/scut/industrial_software/controller/User/ModUsersController.java
@@ -90,7 +90,7 @@ public class ModUsersController {
     @PostMapping("/changePassword")
     public ApiResult<Object> changePassword(@Valid @RequestBody ChangePasswordDTO changePasswordDTO) {
         log.info("用户 {} 正在修改密码", UserHolder.getUser().getId()); // 假设使用 UserHolder 获取当前用户 ID
-        Long userId = UserHolder.getUser().getId();  // 获取当前用户 ID
+        Integer userId = UserHolder.getUser().getId();  // 获取当前用户 ID
 
         return iModUsersService.changePassword(userId, changePasswordDTO);
     }
